fix file not found error
