+++ prowler/config/config.py
@@ -1,3 +1,4 @@
+import os
 import pathlib
 from datetime import datetime, timezone
 from os import getcwd
@@ -5,7 +6,6 @@ from os import getcwd
 import yaml
 
 from prowler.lib.logger import logger
-from prowler.lib.utils.utils import open_file, parse_json_file
 
 timestamp = datetime.today()
 timestamp_utc = datetime.now(timezone.utc).replace(tzinfo=timezone.utc)
@@ -30,8 +30,7 @@ csv_file_suffix = ".csv"
 json_file_suffix = ".json"
 json_asff_file_suffix = ".asff.json"
 html_file_suffix = ".html"
-config_yaml = f"{pathlib.Path().absolute()}/prowler/config/config.yaml"
-print(config_yaml)
+config_yaml = f"{pathlib.Path(os.path.dirname(os.path.realpath(__file__)))}/config.yaml"
 
 
 def change_config_var(variable, value):
@@ -56,22 +55,3 @@ def get_config_var(variable):
     except Exception as error:
         logger.error(f"{error.__class__.__name__}: {error}")
         return ""
-
-
-def get_aws_available_regions():
-    try:
-        actual_directory = pathlib.Path().absolute()
-        f = open_file(
-            f"{actual_directory}/prowler/providers/aws/{aws_services_json_file}"
-        )
-        data = parse_json_file(f)
-
-        regions = set()
-        for service in data["services"].values():
-            for partition in service["regions"]:
-                for item in service["regions"][partition]:
-                    regions.add(item)
-        return list(regions)
-    except Exception as error:
-        logger.error(f"{error.__class__.__name__}: {error}")
-        return []
